separate file structure by domain
