+++ src/main/kotlin/ru/quantick/kinopoisk/movie/controller/MovieController.kt
@@ -1,10 +1,10 @@
-package ru.quantick.kinopoisk.controller
+package ru.quantick.kinopoisk.movie.controller
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
-import ru.quantick.kinopoisk.model.MovieResponse
-import ru.quantick.kinopoisk.model.toResponse
-import ru.quantick.kinopoisk.service.movie.MovieService
+import ru.quantick.kinopoisk.movie.model.MovieResponse
+import ru.quantick.kinopoisk.movie.model.toResponse
+import ru.quantick.kinopoisk.movie.service.MovieService
 
 data class TopMovieResponse(
     val movies: List<MovieResponse>
