+++ src/main/kotlin/ru/quantick/kinopoisk/movie/service/MovieService.kt
@@ -1,0 +1,7 @@
+package ru.quantick.kinopoisk.movie.service
+
+import ru.quantick.kinopoisk.movie.model.Movie
+
+interface MovieService {
+    fun getTopMoviesForTheMonth(): List<Movie>
+}
