+++ src/main/kotlin/ru/quantick/kinopoisk/movie/service/provider/MovieProvider.kt
@@ -1,0 +1,7 @@
+package ru.quantick.kinopoisk.movie.service.provider
+
+import ru.quantick.kinopoisk.movie.model.Movie
+
+interface MovieProvider {
+    fun getTopForTheMonth(): List<Movie>
+}
