--- src/main/kotlin/ru/quantick/kinopoisk/service/movie/MovieService.kt
@@ -1,7 +1,0 @@
-package ru.quantick.kinopoisk.service.movie
-
-import ru.quantick.kinopoisk.model.Movie
-
-interface MovieService {
-    fun getTopMoviesForTheMonth(): List<Movie>
-}
