--- src/main/kotlin/ru/quantick/kinopoisk/service/movie/provider/MovieProvider.kt
@@ -1,7 +1,0 @@
-package ru.quantick.kinopoisk.service.movie.provider
-
-import ru.quantick.kinopoisk.model.Movie
-
-interface MovieProvider {
-    fun getTopForTheMonth(): List<Movie>
-}
